Fix warnings

+++ ext/dba/libinifile/inifile.c
@@ -500,7 +500,7 @@ static int inifile_delete_replace_append(inifile *dba, const key_type *key, cons
 			ret = FAILURE;
 		} else {
 			php_stream_seek(dba->fp, 0, SEEK_END);
-			if (pos_grp_next != php_stream_tell(dba->fp)) {
+			if (pos_grp_next != (size_t)php_stream_tell(dba->fp)) {
 				php_stream_seek(dba->fp, pos_grp_next, SEEK_SET);
 				if (!php_stream_copy_to_stream(dba->fp, fp_tmp, PHP_STREAM_COPY_ALL)) {
 					php_error_docref(NULL TSRMLS_CC, E_WARNING, "Could not copy remainder to temporary stream");
